Revert "Implement basic git commit functionality"
